Вынес форматирование зп в отдельный класс /util/Formatter.php
